[ng] Component refactoring
