+++ firebird-ng/src/app/app.routes.ts
@@ -1,19 +1,12 @@
 import { Routes } from '@angular/router';
-import {MainDisplayComponent} from "./main-display/main-display.component";
-import {FileBrowserComponent} from "./file-browser/file-browser.component";
 import {InputConfigComponent} from "./pages/input-config/input-config.component";
-import {Edm4hepListingComponent} from "./edm4hep-listing/edm4hep-listing.component";
 
 export const routes: Routes = [
   { path: '', redirectTo: '/display', pathMatch: 'full' },
   { path: 'config', component: InputConfigComponent },
   {
-    path: 'files',
-    loadComponent: () => import('./file-browser/file-browser.component').then(m => m.FileBrowserComponent)
-  },
-  {
     path: 'display',
-    loadComponent: () => import('./main-display/main-display.component').then(m => m.MainDisplayComponent)
+    loadComponent: () => import('./pages/main-display/main-display.component').then(m => m.MainDisplayComponent)
   },
   {
     path: 'playground',
@@ -21,10 +14,6 @@ export const routes: Routes = [
   },
   {
     path: 'geometry',
-    loadComponent: () => import('./geometry-tree/geometry-tree.component').then(m => m.GeometryTreeComponent)
-  },
-  {
-    path: 'listing',
-    loadComponent: () => import('./edm4hep-listing/edm4hep-listing.component').then(m => m.Edm4hepListingComponent)
+    loadComponent: () => import('./pages/geometry-tree/geometry-tree.component').then(m => m.GeometryTreeComponent)
   },
 ];
