+++ firebird-ng/src/app/data-pipelines/calorimetry.prettifier.ts
