+++ firebird-ng/src/app/pages/geometry-tree/geometry-tree-window/geometry-tree-window.component.spec.ts
