+++ firebird-ng/src/app/pages/main-display/main-display.component.scss
