+++ firebird-ng/src/app/pages/main-display/main-display.component.spec.ts
