+++ firebird-ng/src/app/pages/main-display/main-display.component.ts
@@ -10,9 +10,9 @@ import {
 import {ClippingSetting, Configuration, PhoenixLoader, PhoenixMenuNode, PresetView} from 'phoenix-event-display';
 import * as THREE from 'three';
 import {Color, DoubleSide, InstancedBufferGeometry, Line, MeshLambertMaterial, MeshPhongMaterial,} from "three";
-import {ALL_GROUPS, GeometryService} from '../services/geometry.service';
+import {ALL_GROUPS, GeometryService} from '../../services/geometry.service';
 import {ActivatedRoute} from '@angular/router';
-import {ThreeGeometryProcessor} from "../data-pipelines/three-geometry.processor";
+import {ThreeGeometryProcessor} from "../../data-pipelines/three-geometry.processor";
 import * as TWEEN from '@tweenjs/tween.js';
 import GUI from "lil-gui";
 import {produceRenderOrder} from "jsrootdi/geom";
@@ -22,18 +22,17 @@ import {
   findObject3DNodes,
   getColorOrDefault,
   pruneEmptyNodes
-} from "../utils/three.utils";
-import {mergeMeshList, MergeResult} from "../utils/three-geometry-merge";
-import {PhoenixThreeFacade} from "../utils/phoenix-three-facade";
+} from "../../utils/three.utils";
+import {PhoenixThreeFacade} from "../../utils/phoenix-three-facade";
 import {BehaviorSubject, Subject} from "rxjs";
-import {GameControllerService} from "../services/game-controller.service";
+import {GameControllerService} from "../../services/game-controller.service";
 import {LineMaterial} from "three/examples/jsm/lines/LineMaterial";
 import {Line2} from "three/examples/jsm/lines/Line2";
 import {LineGeometry} from "three/examples/jsm/lines/LineGeometry";
 import {IoOptionsComponent} from "./io-options/io-options.component";
-import {ProcessTrackInfo, ThreeEventProcessor} from "../data-pipelines/three-event.processor";
-import {UserConfigService} from "../services/user-config.service";
-import {EicAnimationsManager} from "../eic-animation-manager";
+import {ProcessTrackInfo, ThreeEventProcessor} from "../../data-pipelines/three-event.processor";
+import {UserConfigService} from "../../services/user-config.service";
+import {EicAnimationsManager} from "../../phoenix-overload/eic-animation-manager";
 import {MatSlider, MatSliderThumb} from "@angular/material/slider";
 import {MatIcon} from "@angular/material/icon";
 import {MatButton} from "@angular/material/button";
