+++ firebird-ng/src/app/phoenix-overload/eic-animation-manager.ts
